refactor: split models.py into models/ sub-package
The single file had grown to 666 lines (over the project's 500-line max
target). Splitting per chore type keeps each file focused on one state
machine and makes future feature work (e.g. promoting early_window onto
BaseChore) land as localized per-file edits rather than churn across one
monolith.

Public API is unchanged: callers continue to import from .models thanks
to the package __init__.py re-exports. BaseChore.from_dict uses lazy
imports of the subclasses to break the otherwise-circular dependency
between the base and concrete modules.
