+++ custom_components/chore_calendar/models/__init__.py
@@ -1,0 +1,20 @@
+"""Chore data models with state machine logic.
+
+Re-exports the public model classes so callers can keep using
+``from .models import BaseChore, ScheduledChore, IntervalChore, OneshotChore``
+even after the per-type split.
+"""
+
+from __future__ import annotations
+
+from .base import BaseChore
+from .interval import IntervalChore
+from .oneshot import OneshotChore
+from .scheduled import ScheduledChore
+
+__all__ = [
+    "BaseChore",
+    "IntervalChore",
+    "OneshotChore",
+    "ScheduledChore",
+]
